Added changes to save transaction

+++ src/main/java/com/sjsucmpe202/artemis/onlinebankingsystem/entities/accounts/AccountDTO.java
@@ -2,6 +2,7 @@ package com.sjsucmpe202.artemis.onlinebankingsystem.entities.accounts;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -12,7 +13,7 @@ public class AccountDTO {
 
     private Long accountNumber;
     private Date accountOpenDate;
-    private Double accountBalance;
+    private BigDecimal accountBalance;
     private Long accountRoutingNumber;
     private String accountType;
     private String accountInfo;
